+++ ios/Classes/WaveformExtractor.swift
@@ -10,8 +10,10 @@ public class WaveformExtractor {
     var progress: Float = 0.0
     var channelCount: Int = 1
     private var currentProgress: Float = 0.0
-    private let abortWaveformDataQueue = DispatchQueue(label: "WaveformExtractor",
-                                                       attributes: .concurrent)
+    private let abortWaveformDataQueue = DispatchQueue(
+        label: "WaveformExtractor",
+        attributes: .concurrent
+    )
 
     private var _abortGetWaveformData: Bool = false
 
@@ -33,123 +35,166 @@ public class WaveformExtractor {
         audioFile = nil
     }
 
-    public func extractWaveform(samplesPerPixel: Int?,
-                                offset: Int? = 0,
-                                length: UInt? = nil, playerKey: String) -> FloatChannelData?
-    {
+    public func extractWaveform(
+        samplesPerPixel: Int?,
+        offset: Int? = 0,
+        length: UInt? = nil,
+        playerKey: String
+    ) async -> FloatChannelData? {
         guard let audioFile = audioFile else { return nil }
-
-        /// prevent division by zero, + minimum resolution
+        
+        /// Prevent division by zero, + minimum resolution
         let samplesPerPixel = max(1, samplesPerPixel ?? 100)
-
         let currentFrame = audioFile.framePosition
-
-        let totalFrameCount = AVAudioFrameCount(audioFile.length)
-        var framesPerBuffer: AVAudioFrameCount = totalFrameCount / AVAudioFrameCount(samplesPerPixel)
-
-        guard let rmsBuffer = AVAudioPCMBuffer(pcmFormat: audioFile.processingFormat,
-                                               frameCapacity: AVAudioFrameCount(framesPerBuffer)) else { return nil }
-
-        channelCount = Int(audioFile.processingFormat.channelCount)
-        var data = Array(repeating: [Float](zeros: samplesPerPixel), count: channelCount)
-
-        var start: Int
-        if let offset = offset, offset >= 0 {
-            start = offset
-        } else {
-            start = Int(currentFrame / Int64(framesPerBuffer))
-            if let offset = offset, offset < 0 {
-                start += offset
-            }
-
-            if start < 0 {
-                start = 0
-            }
-        }
-        var startFrame: AVAudioFramePosition = offset == nil ? currentFrame : Int64(start * Int(framesPerBuffer))
-
-        var end = samplesPerPixel
-        if let length = length {
-            end = start + Int(length)
-        }
-
-        if end > samplesPerPixel {
-            end = samplesPerPixel
-        }
-        if start > end {
-            result(FlutterError(code: Constants.audioWaveforms, message: "offset is larger than total length.", details: "Please select less number of samples"))
+        let totalFrames = AVAudioFrameCount(audioFile.length)
+        var framesPerBuffer = totalFrames / AVAudioFrameCount(samplesPerPixel)
+        
+        guard let rmsBuffer = AVAudioPCMBuffer(
+            pcmFormat: audioFile.processingFormat,
+            frameCapacity: framesPerBuffer
+        ) else { return nil }
+        
+        let channelCount = Int(audioFile.processingFormat.channelCount)
+        let waveformStorage = WaveformStorage(
+            channelCount: channelCount,
+            size: samplesPerPixel
+        )
+        
+        let startIndex = max(
+            0, offset ?? Int(currentFrame / Int64(framesPerBuffer))
+        )
+        let endIndex = min(
+            samplesPerPixel, startIndex + (length.map { Int($0) } ?? samplesPerPixel)
+        )
+        
+        if startIndex > endIndex {
+            sendErrorToFlutter(
+                message: "Offset is larger than total length.",
+                details: "Please select less number of samples"
+            )
             return nil
         }
-
-        for i in start ..< end {
-
+        
+        var startFrame: AVAudioFramePosition = offset == nil
+        ? currentFrame
+        : Int64(startIndex * Int(framesPerBuffer))
+        
+        for i in startIndex..<endIndex {
             if abortGetWaveformData {
                 audioFile.framePosition = currentFrame
                 abortGetWaveformData = false
                 return nil
             }
-
+            
             do {
                 audioFile.framePosition = startFrame
-                /// Read portion of the buffer
                 try audioFile.read(into: rmsBuffer, frameCount: framesPerBuffer)
-
-            } catch let err as NSError {
-                result(FlutterError(code: Constants.audioWaveforms, message: "Couldn't read into buffer. \(err)", details: nil))
+            } catch {
+                sendErrorToFlutter(
+                    message: "Couldn't read buffer. \(error.localizedDescription)"
+                )
                 return nil
             }
-
+            
             guard let floatData = rmsBuffer.floatChannelData else { return nil }
-            /// Calculating RMS(Root mean square)
-            for channel in 0 ..< channelCount {
-                var rms: Float = 0.0
-                vDSP_rmsqv(floatData[channel], 1, &rms, vDSP_Length(rmsBuffer.frameLength))
-                data[channel][i] = rms
-
+            
+            for channel in 0..<channelCount {
+                /// Calculating RMS(Root mean square)
+                var rmsValue: Float = 0.0
+                vDSP_rmsqv(
+                    floatData[channel], 1, &rmsValue,
+                    vDSP_Length(rmsBuffer.frameLength)
+                )
+                await waveformStorage.update(
+                    channel: channel, index: i, value: rmsValue
+                )
             }
-
-            /// Update progress
-            currentProgress += 1
-            progress = currentProgress / Float(samplesPerPixel)
-
-            /// Send to flutter channel
-            flutterChannel.invokeMethod(Constants.onCurrentExtractedWaveformData, arguments: [
-                Constants.waveformData: getChannelMean(data: data) as Any,
-                Constants.progress: progress,
-                Constants.playerKey: playerKey
-            ])
-
+            
+            let progress = Float(i - startIndex + 1) / Float(endIndex - startIndex)
+            await sendWaveformDataToFlutter(
+                waveformStorage: waveformStorage,
+                progress: progress,
+                playerKey: playerKey
+            )
+            
             startFrame += AVAudioFramePosition(framesPerBuffer)
-
-            if startFrame + AVAudioFramePosition(framesPerBuffer) > totalFrameCount {
-                framesPerBuffer = totalFrameCount - AVAudioFrameCount(startFrame)
+            if startFrame + AVAudioFramePosition(framesPerBuffer) > totalFrames {
+                framesPerBuffer = totalFrames - AVAudioFrameCount(startFrame)
                 if framesPerBuffer <= 0 { break }
             }
         }
-
+        
         audioFile.framePosition = currentFrame
-
-        return data
+        return await waveformStorage.getData()
     }
 
     func getChannelMean(data: FloatChannelData) -> [Float] {
-        waveformData.removeAll()
-        if(channelCount == 2 && data[0].isEmpty == false && data[1].isEmpty == false) {
-            for (ele1, ele2) in zip(data[0], data[1]) {
-                waveformData.append((ele1 + ele2) / 2)
-            }
-        } else if(data[0].isEmpty == false) {
-            waveformData = data[0]
-        }
-        else if (data[1].isEmpty == false) {
-            waveformData = data[1]
+        var resultWaveform = [Float]()
+
+        if channelCount == 2, !data[0].isEmpty, !data[1].isEmpty {
+            resultWaveform = zip(data[0], data[1]).map { ($0 + $1) / 2 }
+        } else if !data[0].isEmpty {
+            resultWaveform = data[0]
+        } else if !data[1].isEmpty {
+            resultWaveform = data[1]
         } else {
-            result(FlutterError(code: Constants.audioWaveforms, message: "Can not get waveform mean", details: "Both audio channels are null"))
+            sendErrorToFlutter(
+                message: "Cannot get waveform mean",
+                details: "Both audio channels are null"
+            )
         }
-        return waveformData
+        return resultWaveform
     }
 
     public func cancel() {
         abortGetWaveformData = true
+    }
+
+    private func sendWaveformDataToFlutter(
+        waveformStorage: WaveformStorage,
+        progress: Float,
+        playerKey: String
+    ) async {
+        let waveformData = await waveformStorage.getData()
+        let meanData = getChannelMean(data: waveformData)
+
+        DispatchQueue.main.async {
+            self.flutterChannel.invokeMethod(
+                Constants.onCurrentExtractedWaveformData,
+                arguments: [
+                    Constants.waveformData: meanData,
+                    Constants.progress: progress,
+                    Constants.playerKey: playerKey
+                ]
+            )
+        }
+    }
+
+    private func sendErrorToFlutter(message: String, details: String? = nil) {
+        DispatchQueue.main.async {
+            self.result(
+                FlutterError(
+                    code: Constants.audioWaveforms,
+                    message: message,
+                    details: details
+                )
+            )
+        }
+    }
+}
+
+actor WaveformStorage {
+    private var data: [[Float]]
+
+    init(channelCount: Int, size: Int) {
+        data = Array(repeating: [Float](repeating: 0, count: size), count: channelCount)
+    }
+
+    func update(channel: Int, index: Int, value: Float) {
+        data[channel][index] = value
+    }
+
+    func getData() -> [[Float]] {
+        return data
     }
 }