+++ lib/src/controllers/player_controller.dart
@@ -191,6 +191,12 @@ class PlayerController extends ChangeNotifier {
     return result;
   }
 
+  /// Stops current waveform extraction, if any.
+  Future<void> stopWaveformExtraction() async {
+    return await AudioWaveformsInterface.instance
+        .stopWaveformExtraction(playerKey);
+  }
+
   /// A function to start the player to play/resume the audio.
   ///
   /// When playing audio is finished, this [player] will be [stopped]